requirements.txt
flask
flask-mail
werkzeug
authlib
requests
python-dotenv
openai
gunicorn
google-genai
google-cloud-storage
flask-babel
google-cloud-translate
